Compute reachability graph over extended BPMN with start/end When computing the reachability graph and N-gram Index, do it over an extended version of the BPMN model that contains two activities per task (one for START another for COMPLETE). In this way, the query for the ongoing state can be done based on START/COMPLETE events, and not assuming atomic activity instances.
When creating the extended BPMN model, two artificial tasks are created per task, one with +START and another one with +COMPLETE at the end. The flow that connects them contains the original ID of the task, so when the N-gram Index returns this flow (activity between start and end), it corresponds to the ID of the activity that is ongoing.